[Trait]: refactor and bugfix
Even TraitDecl can only appear in toplevel in spec, we still move
the entry of parsing to ParseDeclarationSpecifiers just like when
we parse a function definition. BNF grammar does not map to clang
parser very strictly.

For `This` in traitBody methods, we use ThisTy like VoidTy, add some
logic to deal with it like VoidTy does(which was missing before and
caused some segfaults)