[borrowck] A new implementation of BiShengC borrow checker based on NLL(Non Lexical Lifetimes).
It calculates the NLL of region variables by generating a series of
constraints, including liveness constraints, subtyping constraints and
reborrow constraints and solving them. The algorithm has polynomial time
complexity and performs well in complex control flow scenarios.

Note: We provide a DEBUG_PRINT macro to display the execution process of
the borrow checker. To enable it, change DEBUG_PRINT value to 1 in
BSCBorrowChecker.h