[Parser] Fix compiler warning imported from PR 447
The compiler issued a fall-through warning in a switch case during the fix for IC3679.
The default case was already intended to handle the scenario.
To resolve the warning and clarify the control flow, explicitly redirect execution to the default case when the condition does not match `ParseImplTraitDeclaration`.
Additionally, a core dump issue caused by incorrect usage of impl was also fixed.