[source2source]fixes the issue where the const qualifier on function pointer unintended drop during source2source
This change fixes the issue where the const qualifier on function pointer parameters was being dropped during our source-to-source transformations.
We traced the bug to our use of QualType’s normalization routine—which, although intended to keep function signatures consistent, inadvertently stripped the const off parameters. To address it, we’ve added a special-case check up front: whenever we’re doing a source-to-source transform on a function’s input type, we skip the getCanonicalType normalization.
We’ve also revised some of the original test cases that were incorrect and added new tests to cover this scenario.