[clangd] Fix owned-struct destructor TagType SourceLocation
Previously, the auto-inserted destructor for an owned-struct would incorrectly use the SourceLocation of the original owned-struct definition's TagType for its own TagType. This caused clangd to misinterpret the destructor as the definition of the (owned struct S) TagType itself. Consequently, clangd entered a nested infinite loop during TagType resolution, leading to a coredump.

This commit corrects the SourceLocation of the destructor's TagType to prevent this misinterpretation.